feat: define ExpressionMid as an alias
