+++ halo2_middleware/src/circuit.rs
@@ -1,10 +1,8 @@
+use crate::expression::{Expression, Variable};
 use crate::poly::Rotation;
 use crate::{lookup, metadata, permutation, shuffle};
-use core::cmp::max;
-use core::ops::{Add, Mul, Neg, Sub};
 use ff::Field;
 use std::collections::HashMap;
-use std::iter::{Product, Sum};
 
 /// Query of fixed column at a certain relative location
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -64,38 +62,6 @@ pub enum QueryMid {
     Instance(InstanceQueryMid),
 }
 
-pub trait Variable: Clone + Copy + std::fmt::Debug + Eq + PartialEq {
-    /// Degree that an expression would have if it was only this variable.
-    fn degree(&self) -> usize;
-
-    /// Return true if this variable is a virtual abstraction that may be transformed later.  Two
-    /// expressions containing virtual variables cannot be added or multiplied, this means an
-    /// expression can only contain up to one virtual variable.  An expression containing a virtual
-    /// variable cannot appear in a lookup.  Originally this is only fulfilled by the simple
-    /// selector abstraction in the halo2 frontend.
-    /// Overwrite this if the Variable has a virtual case.
-    fn is_virtual(&self) -> bool {
-        false
-    }
-
-    /// Return the name of the virtual variable case.  Originally this is used for "simple
-    /// selector".
-    /// Overwrite this if the Variable has a virtual case.
-    fn virtual_var_case() -> &'static str {
-        unimplemented!();
-    }
-
-    /// Approximate the computational complexity an expression would have if it was only this
-    /// variable.
-    fn complexity(&self) -> usize {
-        0
-    }
-
-    /// Write an identifier of the variable.  If two variables have the same identifier, they must
-    /// be the same variable.
-    fn write_identifier<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()>;
-}
-
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum VarMid {
     /// This is a generic column query
@@ -149,262 +115,13 @@ impl Variable for VarMid {
     }
 }
 
-/// Low-degree expression representing an identity that must hold over the committed columns.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum ExpressionMid<F, V: Variable> {
-    /// This is a constant polynomial
-    Constant(F),
-    /// This is a variable
-    Var(V),
-    /// This is a negated polynomial
-    Negated(Box<ExpressionMid<F, V>>),
-    /// This is the sum of two polynomials
-    Sum(Box<ExpressionMid<F, V>>, Box<ExpressionMid<F, V>>),
-    /// This is the product of two polynomials
-    Product(Box<ExpressionMid<F, V>>, Box<ExpressionMid<F, V>>),
-    /// This is a scaled polynomial
-    Scaled(Box<ExpressionMid<F, V>>, F),
-}
-
-impl<F: Field, V: Variable> ExpressionMid<F, V> {
-    /// Evaluate the polynomial using the provided closures to perform the
-    /// operations.
-    #[allow(clippy::too_many_arguments)]
-    pub fn evaluate<T>(
-        &self,
-        constant: &impl Fn(F) -> T,
-        var: &impl Fn(V) -> T,
-        negated: &impl Fn(T) -> T,
-        sum: &impl Fn(T, T) -> T,
-        product: &impl Fn(T, T) -> T,
-        scaled: &impl Fn(T, F) -> T,
-    ) -> T {
-        match self {
-            ExpressionMid::Constant(scalar) => constant(*scalar),
-            ExpressionMid::Var(v) => var(*v),
-            ExpressionMid::Negated(a) => {
-                let a = a.evaluate(constant, var, negated, sum, product, scaled);
-                negated(a)
-            }
-            ExpressionMid::Sum(a, b) => {
-                let a = a.evaluate(constant, var, negated, sum, product, scaled);
-                let b = b.evaluate(constant, var, negated, sum, product, scaled);
-                sum(a, b)
-            }
-            ExpressionMid::Product(a, b) => {
-                let a = a.evaluate(constant, var, negated, sum, product, scaled);
-                let b = b.evaluate(constant, var, negated, sum, product, scaled);
-                product(a, b)
-            }
-            ExpressionMid::Scaled(a, f) => {
-                let a = a.evaluate(constant, var, negated, sum, product, scaled);
-                scaled(a, *f)
-            }
-        }
-    }
-
-    /// Evaluate the polynomial lazily using the provided closures to perform the
-    /// operations.
-    #[allow(clippy::too_many_arguments)]
-    pub fn evaluate_lazy<T: PartialEq>(
-        &self,
-        constant: &impl Fn(F) -> T,
-        var: &impl Fn(V) -> T,
-        negated: &impl Fn(T) -> T,
-        sum: &impl Fn(T, T) -> T,
-        product: &impl Fn(T, T) -> T,
-        scaled: &impl Fn(T, F) -> T,
-        zero: &T,
-    ) -> T {
-        match self {
-            ExpressionMid::Constant(scalar) => constant(*scalar),
-            ExpressionMid::Var(v) => var(*v),
-            ExpressionMid::Negated(a) => {
-                let a = a.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-                negated(a)
-            }
-            ExpressionMid::Sum(a, b) => {
-                let a = a.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-                let b = b.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-                sum(a, b)
-            }
-            ExpressionMid::Product(a, b) => {
-                let (a, b) = if a.complexity() <= b.complexity() {
-                    (a, b)
-                } else {
-                    (b, a)
-                };
-                let a = a.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-
-                if a == *zero {
-                    a
-                } else {
-                    let b = b.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-                    product(a, b)
-                }
-            }
-            ExpressionMid::Scaled(a, f) => {
-                let a = a.evaluate_lazy(constant, var, negated, sum, product, scaled, zero);
-                scaled(a, *f)
-            }
-        }
-    }
-
-    fn write_identifier<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        match self {
-            ExpressionMid::Constant(scalar) => write!(writer, "{scalar:?}"),
-            ExpressionMid::Var(v) => v.write_identifier(writer),
-            ExpressionMid::Negated(a) => {
-                writer.write_all(b"(-")?;
-                a.write_identifier(writer)?;
-                writer.write_all(b")")
-            }
-            ExpressionMid::Sum(a, b) => {
-                writer.write_all(b"(")?;
-                a.write_identifier(writer)?;
-                writer.write_all(b"+")?;
-                b.write_identifier(writer)?;
-                writer.write_all(b")")
-            }
-            ExpressionMid::Product(a, b) => {
-                writer.write_all(b"(")?;
-                a.write_identifier(writer)?;
-                writer.write_all(b"*")?;
-                b.write_identifier(writer)?;
-                writer.write_all(b")")
-            }
-            ExpressionMid::Scaled(a, f) => {
-                a.write_identifier(writer)?;
-                write!(writer, "*{f:?}")
-            }
-        }
-    }
-
-    /// Identifier for this expression. Expressions with identical identifiers
-    /// do the same calculation (but the expressions don't need to be exactly equal
-    /// in how they are composed e.g. `1 + 2` and `2 + 1` can have the same identifier).
-    pub fn identifier(&self) -> String {
-        let mut cursor = std::io::Cursor::new(Vec::new());
-        self.write_identifier(&mut cursor).unwrap();
-        String::from_utf8(cursor.into_inner()).unwrap()
-    }
-
-    /// Compute the degree of this polynomial
-    pub fn degree(&self) -> usize {
-        use ExpressionMid::*;
-        match self {
-            Constant(_) => 0,
-            Var(v) => v.degree(),
-            Negated(poly) => poly.degree(),
-            Sum(a, b) => max(a.degree(), b.degree()),
-            Product(a, b) => a.degree() + b.degree(),
-            Scaled(poly, _) => poly.degree(),
-        }
-    }
-
-    /// Approximate the computational complexity of this expression.
-    pub fn complexity(&self) -> usize {
-        match self {
-            ExpressionMid::Constant(_) => 0,
-            ExpressionMid::Var(v) => v.complexity(),
-            ExpressionMid::Negated(poly) => poly.complexity() + 5,
-            ExpressionMid::Sum(a, b) => a.complexity() + b.complexity() + 15,
-            ExpressionMid::Product(a, b) => a.complexity() + b.complexity() + 30,
-            ExpressionMid::Scaled(poly, _) => poly.complexity() + 30,
-        }
-    }
-
-    /// Square this expression.
-    pub fn square(self) -> Self {
-        self.clone() * self
-    }
-
-    /// Returns whether or not this expression contains a virtual variable.  Originally this was
-    /// the simple `Selector`.
-    pub fn contains_virtual_var(&self) -> bool {
-        self.evaluate(
-            &|_| false,
-            &|var| var.is_virtual(),
-            &|a| a,
-            &|a, b| a || b,
-            &|a, b| a || b,
-            &|a, _| a,
-        )
-    }
-}
-
-impl<F: Field, V: Variable> Neg for ExpressionMid<F, V> {
-    type Output = ExpressionMid<F, V>;
-    fn neg(self) -> Self::Output {
-        ExpressionMid::Negated(Box::new(self))
-    }
-}
-
-impl<F: Field, V: Variable> Add for ExpressionMid<F, V> {
-    type Output = ExpressionMid<F, V>;
-    fn add(self, rhs: ExpressionMid<F, V>) -> ExpressionMid<F, V> {
-        if self.contains_virtual_var() || rhs.contains_virtual_var() {
-            panic!(
-                "attempted to use a {} in an addition",
-                V::virtual_var_case()
-            );
-        }
-        ExpressionMid::Sum(Box::new(self), Box::new(rhs))
-    }
-}
-
-impl<F: Field, V: Variable> Sub for ExpressionMid<F, V> {
-    type Output = ExpressionMid<F, V>;
-    fn sub(self, rhs: ExpressionMid<F, V>) -> ExpressionMid<F, V> {
-        if self.contains_virtual_var() || rhs.contains_virtual_var() {
-            panic!(
-                "attempted to use a {} in a subtraction",
-                V::virtual_var_case()
-            );
-        }
-        ExpressionMid::Sum(Box::new(self), Box::new(-rhs))
-    }
-}
-
-impl<F: Field, V: Variable> Mul for ExpressionMid<F, V> {
-    type Output = ExpressionMid<F, V>;
-    fn mul(self, rhs: ExpressionMid<F, V>) -> ExpressionMid<F, V> {
-        if self.contains_virtual_var() || rhs.contains_virtual_var() {
-            panic!(
-                "attempted to multiply two expressions containing {}s",
-                V::virtual_var_case()
-            );
-        }
-        ExpressionMid::Product(Box::new(self), Box::new(rhs))
-    }
-}
-
-impl<F: Field, V: Variable> Mul<F> for ExpressionMid<F, V> {
-    type Output = ExpressionMid<F, V>;
-    fn mul(self, rhs: F) -> ExpressionMid<F, V> {
-        ExpressionMid::Scaled(Box::new(self), rhs)
-    }
-}
-
-impl<F: Field, V: Variable> Sum<Self> for ExpressionMid<F, V> {
-    fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.reduce(|acc, x| acc + x)
-            .unwrap_or(ExpressionMid::Constant(F::ZERO))
-    }
-}
-
-impl<F: Field, V: Variable> Product<Self> for ExpressionMid<F, V> {
-    fn product<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.reduce(|acc, x| acc * x)
-            .unwrap_or(ExpressionMid::Constant(F::ONE))
-    }
-}
+pub type ExpressionMid<F> = Expression<F, VarMid>;
 
 /// A Gate contains a single polynomial identity with a name as metadata.
 #[derive(Clone, Debug)]
 pub struct GateV2Backend<F: Field> {
     pub name: String,
-    pub poly: ExpressionMid<F, VarMid>,
+    pub poly: ExpressionMid<F>,
 }
 
 impl<F: Field> GateV2Backend<F> {
@@ -414,7 +131,7 @@ impl<F: Field> GateV2Backend<F> {
     }
 
     /// Returns the polynomial identity of this gate
-    pub fn polynomial(&self) -> &ExpressionMid<F, VarMid> {
+    pub fn polynomial(&self) -> &ExpressionMid<F> {
         &self.poly
     }
 }
@@ -480,7 +197,7 @@ pub trait ColumnType:
     'static + Sized + Copy + std::fmt::Debug + PartialEq + Eq + Into<Any>
 {
     /// Return expression from cell
-    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F, VarMid>;
+    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F>;
 }
 
 /// A column with an index and type
@@ -601,7 +318,7 @@ impl PartialOrd for Any {
 }
 
 impl ColumnType for Advice {
-    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F, VarMid> {
+    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F> {
         ExpressionMid::Var(VarMid::Query(QueryMid::Advice(AdviceQueryMid {
             column_index: index,
             rotation: at,
@@ -610,7 +327,7 @@ impl ColumnType for Advice {
     }
 }
 impl ColumnType for Fixed {
-    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F, VarMid> {
+    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F> {
         ExpressionMid::Var(VarMid::Query(QueryMid::Fixed(FixedQueryMid {
             column_index: index,
             rotation: at,
@@ -618,7 +335,7 @@ impl ColumnType for Fixed {
     }
 }
 impl ColumnType for Instance {
-    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F, VarMid> {
+    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F> {
         ExpressionMid::Var(VarMid::Query(QueryMid::Instance(InstanceQueryMid {
             column_index: index,
             rotation: at,
@@ -626,7 +343,7 @@ impl ColumnType for Instance {
     }
 }
 impl ColumnType for Any {
-    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F, VarMid> {
+    fn query_cell<F: Field>(&self, index: usize, at: Rotation) -> ExpressionMid<F> {
         match self {
             Any::Advice(Advice { phase }) => {
                 ExpressionMid::Var(VarMid::Query(QueryMid::Advice(AdviceQueryMid {
